+++ packages/cosplay-pi-device-hub-client/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log - cosplay-pi-device-hub-client
 
-This log was last generated on Mon, 20 Mar 2023 01:28:53 GMT and should not be manually modified.
+This log was last generated on Mon, 20 Mar 2023 02:45:31 GMT and should not be manually modified.
+
+## 1.1.0
+Mon, 20 Mar 2023 02:45:31 GMT
+
+### Minor changes
+
+- Remove debug log
+- Switch runtime package manager to npm
+- Forward runtime process stdio
 
 ## 1.0.2
 Mon, 20 Mar 2023 01:28:53 GMT